Refactoring and cleanups
